requirements.txt
alembic
click
colorama
Flask
Flask-Cors
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2-binary
python-dateutil
python-editor
python-dotenv
python-jose
rsa
six
SQLAlchemy
SQLAlchemy-Utils
typing-extensions
Werkzeug
zipp